Revert "refactor(qna): 폴더 구조 올바르게 변경 및 service 이름 통일 (#26)"
This reverts commit 3a4a13850a47bfcec924bab38d02ec64b8e28012.
