+++ src/main/java/com/sickgyun/server/core/presentation/qna/dto/QnAResponse.java
@@ -1,4 +1,4 @@
-package com.sickgyun.server.core.qna.presentation.dto;
+package com.sickgyun.server.core.presentation.qna.dto;
 
 import java.time.LocalDateTime;
