+++ src/main/java/com/sickgyun/server/profile/presentation/dto/ProfileCreateRequestDto.java
@@ -7,7 +7,7 @@ import com.sickgyun.server.profile.domain.value.OnlineProfile;
 
 import jakarta.validation.constraints.NotNull;
 
-public record ProfileCreateRequest(
+public record ProfileCreateRequestDto(
 	@NotNull
 	Long userId,
 	String imageUrl,
